Split deployment and fix fixtures
